[REFACTOR][SCRIPT] Move TIRX script printer/builder into src/tirx/script/
Continues the per-dialect TVMScript restructure (commit 2 of 4): the TIRX
printer and ir_builder pieces move from src/script/{printer,ir_builder}/tirx/
into src/tirx/script/{printer,builder}/, joining the rest of the TIRX layer.
Public ir_builder headers move correspondingly to include/tvm/tirx/script/builder/.

The Python parser/builder modules move to python/tvm/tirx/script/, with
backward-compat shim packages left at the old paths so existing imports
(`from tvm.script import tirx as T`, `from tvm.script.ir_builder import tirx`,
`from tvm.script.ir_builder.tirx.utils import buffer_proxy`) keep working.

Notable details:

- src/script/printer/relax/tir.cc (the cross-dialect bridge that prints
  TIRX vars under the relax token) updates its `../tirx/utils.h` include
  to the new `../../../tirx/script/printer/utils.h` location. This file
  stays under src/script/printer/relax/ for now; it will move to
  src/relax/script/printer/ in commit 3.
- src/tirx/script/builder/frame.cc rewrites its
  `../../../tirx/ir/script/script_complete.h` include as
  `../../ir/script/script_complete.h` — same physical file, shorter path
  from the new location.
- A separate shim file `python/tvm/script/ir_builder/tirx/utils.py`
  re-exports `tvm.tirx.script.builder.utils` so that
  `from tvm.script.ir_builder.tirx.utils import buffer_proxy` (used by
  Relax legalization) continues to resolve.
