+++ python/tvm/tirx/script/builder/__init__.py
@@ -1,3 +1,4 @@
+# isort: skip_file
 # Licensed to the Apache Software Foundation (ASF) under one
 # or more contributor license agreements.  See the NOTICE file
 # distributed with this work for additional information
@@ -14,7 +15,9 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
-"""Backward-compat shim for ``tvm.script.parser.tirx`` — moved to ``tvm.tirx.script.parser``."""
+"""Package tvm.script.ir_builder.tirx"""
 
-from tvm.tirx.script.parser import *  # pylint: disable=wildcard-import,redefined-builtin
-from tvm.tirx.script.parser import Buffer, Ptr, macro, prim_func
+from .ir import *  # pylint: disable=wildcard-import,redefined-builtin
+from .ir import boolean as bool  # pylint: disable=redefined-builtin
+from .ir import buffer as Buffer
+from .utils import buffer_proxy, frame_scope, seq_scope
