+++ src/script/printer/relax/tir.cc
@@ -18,7 +18,7 @@
  */
 #include <tvm/ir/expr.h>
 
-#include "../tirx/utils.h"
+#include "../../../tirx/script/printer/utils.h"
 #include "./utils.h"
 
 namespace tvm {
