+++ src/tirx/script/builder/ir.cc
@@ -19,7 +19,7 @@
 #include <tvm/arith/analyzer.h>
 #include <tvm/ffi/container/variant.h>
 #include <tvm/ffi/reflection/registry.h>
-#include <tvm/script/ir_builder/tirx/ir.h>
+#include <tvm/tirx/script/builder/ir.h>
 
 #include "./utils.h"
