+++ src/tirx/script/builder/utils.h
@@ -16,12 +16,12 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-#ifndef TVM_SCRIPT_IR_BUILDER_TIR_UTILS_H_
-#define TVM_SCRIPT_IR_BUILDER_TIR_UTILS_H_
+#ifndef TVM_TIRX_SCRIPT_BUILDER_UTILS_H_
+#define TVM_TIRX_SCRIPT_BUILDER_UTILS_H_
 
-#include <tvm/script/ir_builder/tirx/frame.h>
-#include <tvm/script/ir_builder/tirx/ir.h>
 #include <tvm/tirx/op.h>
+#include <tvm/tirx/script/builder/frame.h>
+#include <tvm/tirx/script/builder/ir.h>
 #include <tvm/tirx/stmt.h>
 
 namespace tvm {
@@ -137,4 +137,4 @@ inline tvm::tirx::BufferRegion BufferRegionFromLoad(tvm::tirx::BufferLoad buffer
 }  // namespace script
 }  // namespace tvm
 
-#endif  // TVM_SCRIPT_IR_BUILDER_TIR_UTILS_H_
+#endif  // TVM_TIRX_SCRIPT_BUILDER_UTILS_H_
