+++ src/tirx/script/printer/buffer.cc
