+++ src/tirx/script/printer/for_loop.cc
