+++ src/tirx/script/printer/function.cc
